requirements.txt
alembic==0.9.8
astroid==1.6.2
backports.functools-lru-cache==1.5
click==6.7
configparser==3.5.0
enum34==1.1.6
feedparser==5.2.1
flake8==3.5.0
Flask==0.12.2
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.8.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
psycopg2==2.7.5
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.8.3
python-dateutil==2.7.0
python-editor==1.0.3
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.5
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
